finished Forecast section
finished following the tutorial for getting the forecast data up and running but I'm confused about the order of the code. within the functions etc. I did not order the code as what is in the tutorial and would like to understand how the order of the code works compared to how I ordered mine.