refactor forecast
rearranged the code for getting the forecast data to match the order of the code from the tutorial. the data is now coming in and i can see a list of forecast data displayed correctly. now going to move onto the third part of the tutorial in order to add more views to the display. need to get this up an running on the watch too.