finished tutorial
finished the tutorial by adding the feature for updated weather data when the watch is shaken. the watch should update data when wrist is shaken and then indicate updated data with vibration. now need to figure out how to get the app on the phone and then put some nice styling to it all.